27 lines in function get_data

+++ libft/ft_min.c
@@ -6,13 +6,13 @@
 /*   By: olbondar <olbondar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/25 17:07:35 by olbondar          #+#    #+#             */
-/*   Updated: 2017/11/25 17:07:49 by olbondar         ###   ########.fr       */
+/*   Updated: 2017/12/02 14:29:55 by olbondar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_min(size_t a, size_t b)
+int	ft_min(int a, int b)
 {
 	if (a < b)
 		return (a);
